Save system implemented
Does it work? I don't know! It is just PlayerPrefs with extra steps? Probably!

Still needs testing to confirm if everything works, or if it's any better than PlayerPrefs. This isn't a very complicated game, so idk if this is needed at all or if it's more efficient in any way shape or form. Tbd.